Create 1e_DataPrep_PM25_EMEP.R
Major updates to the time use PM2.5 processing due to retiring of R packages that handle spatial object, which means a major change in the way R handles spatial objects. Causes code to break. Also updating working directory calls as restructuring the folder. This file only now processes EMEP model, as I have moved the code from a previous file for clarity and brevity. The processing of the AURN and CAMS-Europe are in other files.